% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tempering.R
\name{emcmc}
\alias{emcmc}
\title{Eruptive MCMC}
\usage{
emcmc(f, param = list(x = 0), to.move = TRUE, n = 200,
  cold.block.size = 200, hot.block.size = 100, sd = 1, min.temp = 10,
  max.temp = 100)
}
\arguments{
\item{f}{a function computing log-densities}

\item{param}{a list of named parameters passed on to 'f'; parameters must be given their initial values}

\item{to.move}{a logical or an integer vector indicating which parts of 'param' should be moved in the MCMC}

\item{n}{a number of iterations for the eMCMC}

\item{cold.block.size}{the size of cold blocks}

\item{hot.block.size}{the size of hot blocks}

\item{sd}{the standard deviation of the proposal normal distribution}

\item{min.temp}{the  maximum temperature to be used}

\item{max.temp}{the maximum temperature to be used}
}
\description{
This is a variation on the tempering algorithm. It relies on alternative blocks of cold chains with blocks of very hot chains.
}
\examples{
## try with a basic normal density
f1 <- function(x){dnorm(x, log=TRUE)}

## basic MCMC
mc1 <- metro(f1)
plot(mc1)

## eMCMC
emc1 <- emcmc(f1)
plot(emc1[[1]]) # all chains
plot(emc1[[2]]) # thinned

## try with a mixture of densities
fmix <- function(x){log(dnorm(x, mean=-5) +
                        dnorm(x, mean=5, sd=.5))}

plot(function(x) exp(fmix(x)), xlim=c(-10,10))
title("Target density")

## basic MCMC
mc2 <- metro(fmix, list(x=0), n=6e4)
plot(mc2)

## eMCMC
emc2 <- emcmc(fmix)
plot(emc2[[1]]) # all chains
plot(emc2[[2]]) # thinned
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

