% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mh.R
\name{metro}
\alias{metro}
\title{Basic Metropolis algorithm}
\usage{
metro(f, param = list(x = 0), to.move = TRUE, n = 1000, sd = 1)
}
\arguments{
\item{f}{a function computing log-densities}

\item{param}{a list of named parameters passed on to 'f'; parameters must be given their initial values}

\item{to.move}{a logical or an integer vector indicating which parts of 'param' should be moved in the MCMC}

\item{n}{a number of iterations for the MCMC}

\item{sd}{the standard deviation of the proposal normal distribution}
}
\description{
This is a basic Metropolis algorithm
}
\examples{
## try with a basic normal density
f1 <- function(x){dnorm(x, log=TRUE)}
mc1 <- metro(f1)
head(mc1)
tail(mc1)
plot(mc1)

## change initial point
mc2 <- metro(f1, list(x=-10))
plot(mc2)

## try with a mixture of densities
fmix <- function(x){log(dnorm(x, mean=-3)+dnorm(x, mean=1, sd=.5))}
mc3 <- metro(fmix, list(x=0))
plot(mc3)

## try harder example
fmix2 <- function(x){log(dnorm(x, mean=-5)+dnorm(x, mean=5, sd=.5))}
mc4 <- metro(fmix2, list(x=0), sd=3, n=1e4)
plot(mc4)
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

